feat: add column resource and data source (#9)
Add `kaneo_column` for managing project board columns through the
existing generated client. Existing default columns are imported
explicitly; conflicting creates fail rather than silently adopting them.

- Manage names, icon/color values, final-state flags, and optional
absolute positions; changing the project replaces the column
- Look up columns by project ID and either column ID or stable slug, and
import with `project-id/column-id`
- Update only the managed column’s position and preserve state if a
follow-up ordering request fails
- Retain state when deletion is refused because a column contains tasks,
and confirm missing columns through a successful project column list
- Correct column position types to int32 through the OpenAPI overlay and
regenerate the client
- Add examples, default-column import documentation, unit tests, and
live tests covering all four default imports, conflicting creates, CRUD,
ordering drift, cleared values, imports, task-filled columns, external
deletion, and project replacement

+++ internal/client/client.gen.go
@@ -618,8 +618,8 @@ type Column struct {
 	Name    string `json:"name"`
 
 	// Position Board order, ascending. Columns are always returned sorted by it.
-	Position  float32 `json:"position"`
-	ProjectId string  `json:"projectId"`
+	Position  int32  `json:"position"`
+	ProjectId string `json:"projectId"`
 
 	// Slug Stable identifier derived from the name; tasks store this as their status.
 	Slug      string    `json:"slug"`
@@ -1930,8 +1930,8 @@ type UpdateOrganizationTeamJSONBody struct {
 type ReorderColumnsJSONBody struct {
 	// Columns Every column keeps its new position. Columns from another project are rejected.
 	Columns []struct {
-		Id       string  `json:"id"`
-		Position float32 `json:"position"`
+		Id       string `json:"id"`
+		Position int32  `json:"position"`
 	} `json:"columns"`
 }
